requirements_gemini.txt
# Gemini Ultimate Generator Requirements
# Gemini APIを使用した100点品質レポート生成

# Google AI Python SDK
google-generativeai>=0.3.0

# 環境変数管理
python-dotenv>=1.0.0

# 設定ファイル管理
pyyaml>=6.0

# 既存の依存関係（オプション）
requests>=2.31.0
numpy>=1.24.0